clippy: clear the workspace errors and apply the machine fixes
- Mark the raw-pointer interop entry points unsafe with # Safety docs
  (oak-core upload/download/frame-from-pixels, oak-audio convert) and
  satisfy the existing callers (tests).
- mut_from_ref: allow with the ABI contract documented (the handle
  get_mut helpers in oak-timeline/oak-render/oak-task take the shared
  reference the C ABI passes; exclusivity is the caller's unsafe
  contract).
- Fix the eq_op in the white-balance normalization (green / green).
- Apply cargo clippy --fix across the workspace (redundant closures and
  field names, field reassignment, items after test modules, ...).
- Revert the replace_box fix in image_effect's clip_define: a
  redefinition must allocate a new box, otherwise the old clip handle
  stays valid and the HS-map replace contract (clip != clip2) breaks.
- 283 warnings remain; they are all non-machine-applicable
  (chunks_exact -> as_chunks needs a manual iter_mut, too_many_arguments,
  complex types, missing Safety docs, ...) and are tracked as the
  follow-up.

+++ crates/oak-codec/src/task.rs
@@ -335,8 +335,8 @@ mod tests {
 
 	#[test]
 	fn task_kind_values_match_abi() {
-		assert_eq!(TaskKind::Conform as i32, OAKCODEC_TASK_CONFORM as i32);
-		assert_eq!(TaskKind::Proxy as i32, OAKCODEC_TASK_PROXY as i32);
+		assert_eq!(TaskKind::Conform as i32, OAKCODEC_TASK_CONFORM);
+		assert_eq!(TaskKind::Proxy as i32, OAKCODEC_TASK_PROXY);
 	}
 
 	// ABI constants mirrored from include/codec/task.h.
